requirements.txt
Django==3.2.25
django-ckeditor-5==0.2.17
django-imagekit==5.0.0
pillow==11.2.1
python-dotenv==1.0.0
asgiref==3.8.1
sqlparse==0.5.3 